Fixed errors in the DAO, repository, VM

+++ app/src/main/java/com/example/tutor/repository/ScheduleRepository.kt
@@ -1,5 +1,6 @@
 package com.example.tutor.repository
 
+import android.util.Log
 import androidx.annotation.WorkerThread
 import androidx.lifecycle.LiveData
 import com.example.tutor.bd.dao.ScheduleDAO
@@ -21,7 +22,8 @@ class ScheduleRepository(private val scheduleDAO: ScheduleDAO){
     fun getTotalWeekAmount():Flow<Int> = scheduleDAO.getTotalWeekAmount()
     fun getTotalMonthAmount():Flow<Int> = scheduleDAO.getTotalMonthAmount()
     fun getTotalPeriodAmount(period:String):Flow<Int> = scheduleDAO.getTotalPeriodAmount(period)
-    fun getMapOfPrice(month:String):Map<Long,Int> = scheduleDAO.getMapOfPrice(month)
+    fun getMapOfPrice(month:String):Map<Int,Int> = scheduleDAO.getMapOfPrice(month)
+
 
 
     fun getTotalWeekLessons():Flow<Int> = scheduleDAO.getTotalWeekLessons()
